refactor(payment): PAYPAL-000 moved braintree paypal credit related code to separate folder
